Feature(services-and-our-services-components): initial slider changes
