+++ client/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\App.jsx":"1","C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\Services\\index.jsx":"2","C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\Services\\Services.jsx":"3","C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\OurServices\\OurServices.jsx":"4","C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\OurServices\\inde.jsx":"5"},{"size":155,"mtime":1607388641043,"results":"6","hashOfConfig":"7"},{"size":79,"mtime":1607388612643,"results":"8","hashOfConfig":"7"},{"size":751,"mtime":1607387825475,"results":"9","hashOfConfig":"7"},{"size":1217,"mtime":1607391155808,"results":"10","hashOfConfig":"7"},{"size":40,"mtime":1607389620498,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"151dkl8",{"filePath":"14","messages":"15","errorCount":1,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"source":"16"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\App.jsx",["23","24"],"C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\Services\\index.jsx",["25"],"export { default } from './Services';\r\nexport { default } from './OurServices';","C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\Services\\Services.jsx",[],"C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\OurServices\\OurServices.jsx",["26"],"C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\OurServices\\inde.jsx",[],{"ruleId":"27","severity":1,"message":"28","line":2,"column":3,"nodeType":"29","messageId":"30","endLine":2,"endColumn":11},{"ruleId":"27","severity":1,"message":"31","line":2,"column":13,"nodeType":"29","messageId":"30","endLine":2,"endColumn":21},{"ruleId":null,"fatal":true,"severity":2,"message":"32","line":2,"column":10},{"ruleId":"27","severity":1,"message":"33","line":1,"column":17,"nodeType":"29","messageId":"30","endLine":1,"endColumn":26},"no-unused-vars","'PostList' is defined but never used.","Identifier","unusedVar","'Services' is defined but never used.","Parsing error: Only one default export allowed per module.\n\n  1 | export { default } from './Services';\n> 2 | export { default } from './OurServices';\n    |          ^","'Component' is defined but never used."]
+[{"C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\App.jsx":"1","C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\Services\\index.jsx":"2","C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\Services\\Services.jsx":"3","C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\OurServices\\OurServices.jsx":"4","C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\OurServices\\inde.jsx":"5","C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\index.jsx":"6","C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\PostList\\PostList.jsx":"7","C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\PostList\\index.jsx":"8"},{"size":155,"mtime":1607388641043,"results":"9","hashOfConfig":"10"},{"size":79,"mtime":1607388612643,"results":"11","hashOfConfig":"10"},{"size":388,"mtime":1607432864000,"results":"12","hashOfConfig":"10"},{"size":1783,"mtime":1607450323809,"results":"13","hashOfConfig":"10"},{"size":40,"mtime":1607389620498,"results":"14","hashOfConfig":"10"},{"size":164,"mtime":1607389620498,"results":"15","hashOfConfig":"10"},{"size":661,"mtime":1607432840665,"results":"16","hashOfConfig":"10"},{"size":39,"mtime":1607387375953,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"151dkl8",{"filePath":"20","messages":"21","errorCount":1,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"source":"22"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\App.jsx",["35","36"],"C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\Services\\index.jsx",["37"],"export { default } from './Services';\r\nexport { default } from './OurServices';","C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\Services\\Services.jsx",[],"C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\OurServices\\OurServices.jsx",[],"C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\OurServices\\inde.jsx",[],"C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\index.jsx",[],"C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\PostList\\PostList.jsx",[],"C:\\Users\\Daniel\\Desktop\\CITi\\PTA\\despair\\Delta-odonto\\client\\src\\components\\PostList\\index.jsx",[],{"ruleId":"38","severity":1,"message":"39","line":2,"column":3,"nodeType":"40","messageId":"41","endLine":2,"endColumn":11},{"ruleId":"38","severity":1,"message":"42","line":2,"column":13,"nodeType":"40","messageId":"41","endLine":2,"endColumn":21},{"ruleId":null,"fatal":true,"severity":2,"message":"43","line":2,"column":10},"no-unused-vars","'PostList' is defined but never used.","Identifier","unusedVar","'Services' is defined but never used.","Parsing error: Only one default export allowed per module.\n\n  1 | export { default } from './Services';\n> 2 | export { default } from './OurServices';\n    |          ^"]
