+++ client/src/components/OurServices/OurServices.css
@@ -1,4 +1,5 @@
 .ourServicesSection {
   display: flex;
   flex-direction: column;
+  background-color:gray;
 }
