+++ client/src/components/Services/Services.css
@@ -5,15 +5,21 @@
   padding: 0;
 }
 
+.services-list {
+  display: flex;
+}
+
 .service-card {
   margin-top: 1vh;
   display: flex;
   flex-direction: column;
-  width: 66.5vw;
+  width: 64vh;
   height: 63vh;
   position: relative;
   overflow: hidden;
-  background-color: red;
+  background-color: darkgray;
+  align-items: center;
+  background: linear-gradient(180deg, rgba(3, 44, 102, 0.0104167) 1.51%, rgba(3, 44, 102, 0) 24.48%, rgba(3, 44, 102, 0.5) 72.32%);
 }
 
 .service-card h1 {
@@ -23,12 +29,13 @@
   font-weight: 600;
   font-size: 95%;
   position: absolute;
-  top: 40vh;
-  margin-left: 2vw;
-  color: #FFFFFF;
+  top: 55vh;
+  /* margin-left: 2vw; */
+  color: #FFFFFF;  
 }
 
 .service-card p {
+  display: none;
   z-index: 1;
   font-family: Poppins;
   font-style: normal;
@@ -50,6 +57,5 @@
   -webkit-transform: translate(-50%,-50%);
     -ms-transform: translate(-50%,-50%);
         transform: translate(-50%,-50%); 
-  background: linear-gradient(180deg, rgba(3, 44, 102, 0.0104167) 1.51%, rgba(3, 44, 102, 0) 24.48%, rgba(3, 44, 102, 0.5) 72.32%);
 }
