+++ client/src/components/Services/Services.jsx
@@ -1,29 +1,16 @@
-import { useState, useEffect } from 'react';
-import axios from 'axios';
+import React from 'react';
 
 import './Services.css';
 
-function Services() {
-  const [services, setServices] = useState([]);
-
-  const loadServices = async () => {
-    const res = await axios.get('http://localhost:3001/api/services');
-    setServices(res.data);
-  };
-
-  useEffect(() => {
-    loadServices();
-  }, []);
-
+function Services(props) {
+  
   return (
-    <div className="services-list">
-      {services?.map(({ _id, title, image, description }) => (
-        <div key={_id} className="service-card">
-          <h1>{title}</h1>
-          <p>{description}</p>
-          <img src={image[0]?.url} alt="Imagem do serviço"/>          
-        </div>
-      ))}
+    <div className="services-list">  
+      <div className="service-card">
+        <h1>{props.title}</h1>
+        <p>{props.description}</p>
+        <img src={props.image[0]?.url} alt="Imagem do serviço"/>          
+      </div> 
     </div>
   );
 }
